Refactor to abstract VM used to run Lua code
